Process links inside html2dita, and convert their hrefs accordingly
